requirements.txt
langchain-openai
python-dotenv
langchain
langchain-community
langchain-ollama
langchain-core
streamlit